Smart_Water_Filter_Alert/requirements.txt
matplotlib
paho-mqtt
PyQt5
icecream